Refactor: Move clear and scroll functionalities to separate modules with tests
